Spit the clients into separate modules to eliminate the boto3 dependency for users who only need the SimpleOsduClient.
