+++ requirements.txt
@@ -1,2 +1,1 @@
-boto3>=1.14
-requests>=2.0
+requests==2.20.*
